+++ resources/js/components/utils/firestore/index.js
@@ -4,10 +4,10 @@ import 'firebase/firestore'
 import 'firebase/database'
 import 'firebase/storage'
 
-const config =
+const config = 
     process.env.NODE_ENV === 'development' ?
     JSON.parse(process.env.VUE_APP_FIREBASE_CONFIG) :
-    JSON.parse(process.env.VUE_APP_FIREBASE_CONFIG_PUBLIC)
+    JSON.parse(process.env.VUE_APP_FIREBASE_CONFIG_PUBLIC);
 
 firebaseApp.initializeApp(config)