+++ resources/lang/en/badges.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+
+    'add' => 'Add badge',
+    'create' => 'Create badge',
+    'edit' => 'Edit badge',
+    'icon_select' => 'Select icon',
+    'icon' => 'Icon',
+    'fontawesome' => 'Do you want more icons? Visit: ',
+    'title' => 'Title',
+    'description' => 'Description',
+    'attributes' => 'Atributes',
+    'attributes_info' => 'positive values for good behaviours, negatives for the bad ones',
+
+    'hp' => 'Health points',
+    'experience' => 'Experience',
+    'gold' => 'Gold',
+];